plugins/vecto_metadata/requirements.txt
Nikola
doit
logbook
gitpython
json
